+++ herd-view/src/app/bin/auth/auth.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { map, take } from 'rxjs/operators';
 import { AuthService } from './auth.service';
@@ -9,8 +9,7 @@ import { AuthService } from './auth.service';
 })
 export class AuthGuard implements CanActivate {
   constructor(
-    private authService: AuthService,
-    private router: Router
+    private authService: AuthService
   ) {}
 
   canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
@@ -19,11 +18,10 @@ export class AuthGuard implements CanActivate {
         take(1),
         map((isLoggedIn: boolean) => {
           if (!isLoggedIn) {
-            this.router.navigate(['/login']);
             return false;
           }
           return true;
         })
-      )
+      );
   }
 }