+++ herd-view/src/app/services/base/redherd-api.ts
@@ -3,6 +3,7 @@ import { Observable, throwError } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
 import { RedHerdWebContext } from './redherd-web';
 import { RedHerdObject, RedHerdEntity, RedHerdRootEndpoint } from '../../bin/model/base/redherd-common';
+import { JwtToken } from '../../bin/model/token';
 import { JSendResponse, JSendResponseInspector } from '../../bin/model/jsend';
 import { Config } from '../../config';
 import { NotificationsService } from 'angular2-notifications';
@@ -32,14 +33,14 @@ export abstract class RedHerdApi {
     this.context = new RedHerdApiContext(endpoint);
   }
 
-  public get Token(): string {
+  public get Token(): JwtToken {
     return this.context.Token;
   }
 
   /** GET: retrieve objects from server */
   protected getAll<T>(entity: RedHerdEntity): Observable<T[]> {
     //const url = `${this.context.Url}`;
-    const url = this.context.Token ? `${this.context.Url}?t=${this.context.Token}` : `${this.context.Url}`;
+    const url = !this.context.Token.isEmpty() ? `${this.context.Url}?t=${this.context.Token.value}` : `${this.context.Url}`;
     const options = this.context.HttpOptions;
 
     return this.http.get<JSendResponse>(url, options)
@@ -58,7 +59,7 @@ export abstract class RedHerdApi {
   /** GET: retrieve object by id */
   protected get<T>(entity: RedHerdEntity, param: number | string): Observable<T> {
     //const url = `${this.context.Url}/${param}`;
-    const url = this.context.Token ? `${this.context.Url}/${param}?t=${this.context.Token}` : `${this.context.Url}/${param}`;
+    const url = !this.context.Token.isEmpty() ? `${this.context.Url}/${param}?t=${this.context.Token.value}` : `${this.context.Url}/${param}`;
     const options = this.context.HttpOptions;
 
     return this.http.get<JSendResponse>(url, options)
@@ -77,7 +78,7 @@ export abstract class RedHerdApi {
   /** POST: add new object to server */
   protected add(object: RedHerdObject): Observable<boolean> {
     //const url = `${this.context.Url}`;
-    const url = this.context.Token ? `${this.context.Url}?t=${this.context.Token}` : `${this.context.Url}`;
+    const url = !this.context.Token.isEmpty() ? `${this.context.Url}?t=${this.context.Token.value}` : `${this.context.Url}`;
     const options = this.context.HttpOptions;
 
     return this.http.post<JSendResponse>(url, object, options)
@@ -96,7 +97,7 @@ export abstract class RedHerdApi {
   /** DELETE: delete an object from server */
   protected delete(id: number): Observable<boolean> {
     //const url = `${this.context.Url}/${id}`;
-    const url = this.context.Token ? `${this.context.Url}/${id}?t=${this.context.Token}` : `${this.context.Url}/${id}`;
+    const url = !this.context.Token.isEmpty() ? `${this.context.Url}/${id}?t=${this.context.Token.value}` : `${this.context.Url}/${id}`;
     const options = this.context.HttpOptions;
 
     return this.http.delete<JSendResponse>(url, options)
@@ -115,7 +116,7 @@ export abstract class RedHerdApi {
   /** PUT: update an object on server */
   protected update(object: RedHerdObject): Observable<boolean> {
     //const url = `${this.context.Url}/${object.id}`;
-    const url = this.context.Token ? `${this.context.Url}/${object.id}?t=${this.context.Token}` : `${this.context.Url}/${object.id}`;
+    const url = !this.context.Token.isEmpty() ? `${this.context.Url}/${object.id}?t=${this.context.Token.value}` : `${this.context.Url}/${object.id}`;
     const options = this.context.HttpOptions;
 
     return this.http.put<JSendResponse>(url, object, options)