+++ herd-view/src/app/services/base/redherd-web.ts
@@ -1,9 +1,11 @@
 import { HttpHeaders } from '@angular/common/http';
+import { JwtToken } from '../../bin/model/token';
 import { Config } from '../../config';
 
 
 export class RedHerdWebContext {
   protected httpOptions: { headers: HttpHeaders };
+  protected token: JwtToken;
   
   constructor() {
     this.httpOptions = { 
@@ -13,11 +15,10 @@ export class RedHerdWebContext {
     };
   }
 
-  public get Token(): string {
-    // Switched to sessionStorage API in order to mitigate the effect of the
-    // browser dirty cache after multiple framework deploy
-    //return localStorage.getItem(Config.auth_token_store);
-    return sessionStorage.getItem(Config.auth_token_store);
+  public get Token(): JwtToken {
+    return new JwtToken(
+      Config.single_instance ? sessionStorage.getItem(Config.auth_token_store) : localStorage.getItem(Config.auth_token_store)
+    );
   }
 
   public get HttpOptions(): { headers: HttpHeaders } {